% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.geonetwork}
\alias{plot.geonetwork}
\title{Plot a geonetwork}
\usage{
\method{plot}{geonetwork}(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{geonetwork}.}

\item{y}{Ignored.}

\item{...}{Further specifications passed to \link{plot_sf}.}
}
\description{
Plot one or more attributes of a geonetwork on a map
}
\examples{
plot(eurodist, axes = TRUE, type = "n")
plot(sf::st_geometry(spData::world), col = "lightgray", add = TRUE)
plot(eurodist, axes = TRUE, add = TRUE)
}
