% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eurodist}
\alias{eurodist}
\title{Distances between 21 European cities}
\format{\code{geonetwork}.}
\source{
Distances (in km) between 21 European cities are taken
  from \code{\link[datasets]{eurodist}}. Coordinates of the cities
  were obtained with \code{\link[ggmap]{geocode}}.
}
\usage{
eurodist
}
\description{
A \code{geonetwork} object representing a full graph of 21
European cities with edges weighted by distance in km.
}
\keyword{datasets}
